+++ spec/features/merge_request/user_sees_discussions_spec.rb
@@ -88,5 +88,13 @@ describe 'Merge request > User sees discussions', :js do
         expect(page).to have_content "started a discussion on commit #{note.commit_id[0...7]}"
       end
     end
+
+    context 'a commit non-diff discussion' do
+      let(:note) { create(:discussion_note_on_commit, project: project) }
+
+      it 'displays correct header' do
+        expect(page).to have_content "commented on commit #{note.commit_id[0...7]}"
+      end
+    end
   end
 end