+++ spec/javascripts/fixtures/jobs.rb
@@ -34,21 +34,22 @@ describe Projects::JobsController, '(JavaScript fixtures)', type: :controller do
   end
 
   it 'builds/build-with-artifacts.html.raw' do |example|
-    get :show,
+    get :show, params: {
       namespace_id: project.namespace.to_param,
       project_id: project,
       id: build_with_artifacts.to_param
+    }
 
     expect(response).to be_success
     store_frontend_fixture(response, example.description)
   end
 
   it 'jobs/delayed.json' do |example|
-    get :show,
-        namespace_id: project.namespace.to_param,
-        project_id: project,
-        id: delayed_job.to_param,
-        format: :json
+    get :show, params: {
+      namespace_id: project.namespace.to_param,
+      project_id: project,
+      id: delayed_job.to_param
+    }, format: :json
 
     expect(response).to be_success
     store_frontend_fixture(response, example.description)