+++ spec/models/broadcast_message_spec.rb
@@ -49,7 +49,7 @@ describe BroadcastMessage do
     it 'caches the output of the query' do
       create(:broadcast_message)
 
-      expect(described_class).to receive(:where).and_call_original.once
+      expect(described_class).to receive(:current_and_future_messages).and_call_original.once
 
       described_class.current
 
@@ -92,27 +92,6 @@ describe BroadcastMessage do
 
       expect(Rails.cache).to receive(:delete).with(described_class::LEGACY_CACHE_KEY)
       expect(described_class.current.length).to eq(0)
-    end
-
-    it 'gracefully handles bad cache entry' do
-      allow(described_class).to receive(:current_and_future_messages).and_return('{')
-
-      expect(described_class.current).to be_empty
-    end
-
-    it 'gracefully handles an empty hash' do
-      allow(described_class).to receive(:current_and_future_messages).and_return('{}')
-
-      expect(described_class.current).to be_empty
-    end
-
-    it 'gracefully handles unknown attributes' do
-      message = create(:broadcast_message)
-
-      allow(described_class).to receive(:current_and_future_messages)
-                                 .and_return([{ bad_attr: 1 }, message])
-
-      expect(described_class.current).to eq([message])
     end
   end